feat(disclosure.utils): Move has_been_extracted/pdfed to utils
